+++ app.py
@@ -641,438 +641,6 @@ def get_category_id(category_name, description=None, user_id=None):
     return None
 
 
-def create_default_category_mappings(user_id):
-    """Create default category mappings for a new user"""
-    # Check if user already has any mappings
-    existing_mappings_count = CategoryMapping.query.filter_by(user_id=user_id).count()
-    
-    # Only create defaults if user has no mappings
-    if existing_mappings_count > 0:
-        return
-    
-    # Get user's categories to map to
-    # We'll need to find the appropriate category IDs for the current user
-    categories = {}
-    
-    # Find common top-level categories
-    for category_name in ["Food", "Transportation", "Housing", "Shopping", "Entertainment", "Health", "Personal", "Other"]:
-        category = Category.query.filter_by(
-            user_id=user_id,
-            name=category_name,
-            parent_id=None
-        ).first()
-        
-        if category:
-            categories[category_name.lower()] = category.id
-            
-            # Also get subcategories
-            for subcategory in category.subcategories:
-                categories[subcategory.name.lower()] = subcategory.id
-    
-    # If we couldn't find any categories, we can't create mappings
-    if not categories:
-        app.logger.warning(f"Could not create default category mappings for user {user_id}: no categories found")
-        return
-    
-    # Default mappings as (keyword, category_key, is_regex, priority)
-    default_mappings = [
-        # Food & Dining
-        ("grocery", "groceries", False, 5),
-        ("groceries", "groceries", False, 5),
-        ("supermarket", "groceries", False, 5),
-        ("walmart", "groceries", False, 3),
-        ("target", "groceries", False, 3),
-        ("costco", "groceries", False, 5),
-        ("safeway", "groceries", False, 5),
-        ("kroger", "groceries", False, 5),
-        ("aldi", "groceries", False, 5),
-        ("trader joe", "groceries", False, 5),
-        ("whole foods", "groceries", False, 5),
-        ("wegmans", "groceries", False, 5),
-        ("publix", "groceries", False, 5),
-        ("sprouts", "groceries", False, 5),
-        ("sams club", "groceries", False, 5),
-
-        # Restaurants
-        ("restaurant", "restaurants", False, 5),
-        ("dining", "restaurants", False, 5),
-        ("takeout", "restaurants", False, 5),
-        ("doordash", "restaurants", False, 5),
-        ("ubereats", "restaurants", False, 5),
-        ("grubhub", "restaurants", False, 5),
-        ("mcdonald", "restaurants", False, 5),
-        ("burger", "restaurants", False, 4),
-        ("pizza", "restaurants", False, 4),
-        ("chipotle", "restaurants", False, 5),
-        ("panera", "restaurants", False, 5),
-        ("kfc", "restaurants", False, 5),
-        ("wendy's", "restaurants", False, 5),
-        ("taco bell", "restaurants", False, 5),
-        ("chick-fil-a", "restaurants", False, 5),
-        ("five guys", "restaurants", False, 5),
-        ("ihop", "restaurants", False, 5),
-        ("denny's", "restaurants", False, 5),
-
-        # Coffee shops
-        ("starbucks", "coffee shops", False, 5),
-        ("coffee", "coffee shops", False, 4),
-        ("dunkin", "coffee shops", False, 5),
-        ("peet", "coffee shops", False, 5),
-        ("tim hortons", "coffee shops", False, 5),
-
-        # Gas & Transportation
-        ("gas station", "gas", False, 5),
-        ("gasoline", "gas", False, 5),
-        ("fuel", "gas", False, 5),
-        ("chevron", "gas", False, 5),
-        ("shell", "gas", False, 5),
-        ("exxon", "gas", False, 5),
-        ("tesla supercharger", "gas", False, 5),
-        ("ev charging", "gas", False, 5),
-
-        # Rideshare & Transit
-        ("uber", "rideshare", False, 5),
-        ("lyft", "rideshare", False, 5),
-        ("taxi", "rideshare", False, 5),
-        ("transit", "public transit", False, 5),
-        ("subway", "public transit", False, 5),
-        ("bus", "public transit", False, 5),
-        ("train", "public transit", False, 5),
-        ("amtrak", "public transit", False, 5),
-        ("greyhound", "public transit", False, 5),
-        ("parking", "transportation", False, 5),
-        ("toll", "transportation", False, 5),
-        ("bike share", "transportation", False, 5),
-        ("scooter rental", "transportation", False, 5),
-
-        # Housing & Utilities
-        ("rent", "rent/mortgage", False, 5),
-        ("mortgage", "rent/mortgage", False, 5),
-        ("airbnb", "rent/mortgage", False, 5),
-        ("vrbo", "rent/mortgage", False, 5),
-        ("water bill", "utilities", False, 5),
-        ("electric", "utilities", False, 5),
-        ("utility", "utilities", False, 5),
-        ("utilities", "utilities", False, 5),
-        ("internet", "utilities", False, 5),
-        ("Ngrid", "utilities", False, 5),
-        ("maintenance", "home maintenance", False, 4),
-        ("repair", "home maintenance", False, 4),
-        ("hvac", "home maintenance", False, 5),
-        ("pest control", "home maintenance", False, 5),
-        ("home security", "home maintenance", False, 5),
-        ("home depot", "home maintenance", False, 5),
-        ("lowe's", "home maintenance", False, 5),
-
-        # Shopping
-        ("amazon", "shopping", False, 5),
-        ("ebay", "shopping", False, 5),
-        ("etsy", "shopping", False, 5),
-        ("clothing", "clothing", False, 5),
-        ("apparel", "clothing", False, 5),
-        ("shoes", "clothing", False, 5),
-        ("electronics", "electronics", False, 5),
-        ("best buy", "electronics", False, 5),
-        ("apple", "electronics", False, 5),
-        ("microsoft", "electronics", False, 5),
-        ("furniture", "shopping", False, 5),
-        ("homegoods", "shopping", False, 5),
-        ("ikea", "shopping", False, 5),
-        ("tj maxx", "shopping", False, 5),
-        ("marshalls", "shopping", False, 5),
-        ("nordstrom", "shopping", False, 5),
-        ("macys", "shopping", False, 5),
-        ("zara", "shopping", False, 5),
-        ("uniqlo", "shopping", False, 5),
-        ("shein", "shopping", False, 5),
-
-        # Entertainment & Subscriptions
-        ("movie", "movies", False, 5),
-        ("cinema", "movies", False, 5),
-        ("theater", "movies", False, 5),
-        ("amc", "movies", False, 5),
-        ("regal", "movies", False, 5),
-        ("netflix", "subscriptions", False, 5),
-        ("hulu", "subscriptions", False, 5),
-        ("spotify", "subscriptions", False, 5),
-        ("apple music", "subscriptions", False, 5),
-        ("disney+", "subscriptions", False, 5),
-        ("hbo", "subscriptions", False, 5),
-        ("prime video", "subscriptions", False, 5),
-        ("paramount+", "subscriptions", False, 5),
-        ("game", "entertainment", False, 4),
-        ("playstation", "entertainment", False, 5),
-        ("xbox", "entertainment", False, 5),
-        ("nintendo", "entertainment", False, 5),
-        ("concert", "entertainment", False, 5),
-        ("festival", "entertainment", False, 5),
-        ("sports ticket", "entertainment", False, 5),
-
-        # Health & Wellness
-        ("gym", "health", False, 5),
-        ("fitness", "health", False, 5),
-        ("doctor", "health", False, 5),
-        ("dentist", "health", False, 5),
-        ("hospital", "health", False, 5),
-        ("pharmacy", "health", False, 5),
-        ("walgreens", "health", False, 5),
-        ("cvs", "health", False, 5),
-        ("rite aid", "health", False, 5),
-        ("vision", "health", False, 5),
-        ("glasses", "health", False, 5),
-        ("contacts", "health", False, 5),
-        ("insurance", "health", False, 5),
-    ]
-
-    
-    # Create the mappings
-    for keyword, category_key, is_regex, priority in default_mappings:
-        # Check if we have a matching category for this keyword
-        if category_key in categories:
-            category_id = categories[category_key]
-            
-            # Create the mapping
-            mapping = CategoryMapping(
-                user_id=user_id,
-                keyword=keyword,
-                category_id=category_id,
-                is_regex=is_regex,
-                priority=priority,
-                match_count=0,
-                active=True
-            )
-            
-            db.session.add(mapping)
-    
-    # Commit all mappings at once
-    try:
-        db.session.commit()
-        app.logger.info(f"Created default category mappings for user {user_id}")
-    except Exception as e:
-        db.session.rollback()
-        app.logger.error(f"Error creating default category mappings: {str(e)}")
-
-# Then modify the existing create_default_categories function to also create mappings:
-
-def create_default_categories(user_id):
-    """Create default expense categories for a new user"""
-    default_categories = [
-        # Housing
-        {"name": "Housing", "icon": "fa-home", "color": "#3498db", "subcategories": [
-            {"name": "Rent/Mortgage", "icon": "fa-building", "color": "#3498db"},
-            {"name": "Utilities", "icon": "fa-bolt", "color": "#3498db"},
-            {"name": "Home Maintenance", "icon": "fa-tools", "color": "#3498db"}
-        ]},
-        # Food
-        {"name": "Food", "icon": "fa-utensils", "color": "#e74c3c", "subcategories": [
-            {"name": "Groceries", "icon": "fa-shopping-basket", "color": "#e74c3c"},
-            {"name": "Restaurants", "icon": "fa-hamburger", "color": "#e74c3c"},
-            {"name": "Coffee Shops", "icon": "fa-coffee", "color": "#e74c3c"}
-        ]},
-        # Transportation
-        {"name": "Transportation", "icon": "fa-car", "color": "#2ecc71", "subcategories": [
-            {"name": "Gas", "icon": "fa-gas-pump", "color": "#2ecc71"},
-            {"name": "Public Transit", "icon": "fa-bus", "color": "#2ecc71"},
-            {"name": "Rideshare", "icon": "fa-taxi", "color": "#2ecc71"}
-        ]},
-        # Shopping
-        {"name": "Shopping", "icon": "fa-shopping-cart", "color": "#9b59b6", "subcategories": [
-            {"name": "Clothing", "icon": "fa-tshirt", "color": "#9b59b6"},
-            {"name": "Electronics", "icon": "fa-laptop", "color": "#9b59b6"},
-            {"name": "Gifts", "icon": "fa-gift", "color": "#9b59b6"}
-        ]},
-        # Entertainment
-        {"name": "Entertainment", "icon": "fa-film", "color": "#f39c12", "subcategories": [
-            {"name": "Movies", "icon": "fa-ticket-alt", "color": "#f39c12"},
-            {"name": "Music", "icon": "fa-music", "color": "#f39c12"},
-            {"name": "Subscriptions", "icon": "fa-play-circle", "color": "#f39c12"}
-        ]},
-        # Health
-        {"name": "Health", "icon": "fa-heartbeat", "color": "#1abc9c", "subcategories": [
-            {"name": "Medical", "icon": "fa-stethoscope", "color": "#1abc9c"},
-            {"name": "Pharmacy", "icon": "fa-prescription-bottle", "color": "#1abc9c"},
-            {"name": "Fitness", "icon": "fa-dumbbell", "color": "#1abc9c"}
-        ]},
-        # Personal
-        {"name": "Personal", "icon": "fa-user", "color": "#34495e", "subcategories": [
-            {"name": "Self-care", "icon": "fa-spa", "color": "#34495e"},
-            {"name": "Education", "icon": "fa-graduation-cap", "color": "#34495e"}
-        ]},
-        # Other
-        {"name": "Other", "icon": "fa-question-circle", "color": "#95a5a6", "is_system": True}
-    ]
-
-    for cat_data in default_categories:
-        subcategories = cat_data.pop('subcategories', [])
-        category = Category(user_id=user_id, **cat_data)
-        db.session.add(category)
-        db.session.flush()  # Get the ID without committing
-
-        for subcat_data in subcategories:
-            subcat = Category(user_id=user_id, parent_id=category.id, **subcat_data)
-            db.session.add(subcat)
-
-    db.session.commit()
-    
-    # Create default category mappings after creating categories
-    create_default_category_mappings(user_id)
-
-def create_default_budgets(user_id):
-    """Create default budget templates for a new user, all deactivated by default"""
-    from app import Budget, Category, db
-
-    # Get the user's categories first
-    categories = Category.query.filter_by(user_id=user_id).all()
-    category_map = {}
-    
-    # Create a map of category types to their IDs
-    for category in categories:
-        if category.name == "Housing":
-            category_map['housing'] = category.id
-        elif category.name == "Food":
-            category_map['food'] = category.id
-        elif category.name == "Transportation":
-            category_map['transportation'] = category.id
-        elif category.name == "Entertainment":
-            category_map['entertainment'] = category.id
-        elif category.name == "Shopping":
-            category_map['shopping'] = category.id
-        elif category.name == "Health":
-            category_map['health'] = category.id
-        elif category.name == "Personal":
-            category_map['personal'] = category.id
-        elif category.name == "Other":
-            category_map['other'] = category.id
-    
-    # Default budget templates with realistic amounts
-    default_budgets = [
-        {
-            'name': 'Monthly Housing Budget',
-            'category_type': 'housing',
-            'amount': 1200,
-            'period': 'monthly',
-            'include_subcategories': True
-        },
-        {
-            'name': 'Monthly Food Budget',
-            'category_type': 'food',
-            'amount': 600,
-            'period': 'monthly',
-            'include_subcategories': True
-        },
-        {
-            'name': 'Monthly Transportation',
-            'category_type': 'transportation',
-            'amount': 400,
-            'period': 'monthly',
-            'include_subcategories': True
-        },
-        {
-            'name': 'Monthly Entertainment',
-            'category_type': 'entertainment',
-            'amount': 200,
-            'period': 'monthly',
-            'include_subcategories': True
-        },
-        {
-            'name': 'Monthly Shopping',
-            'category_type': 'shopping',
-            'amount': 300,
-            'period': 'monthly',
-            'include_subcategories': True
-        },
-        {
-            'name': 'Monthly Healthcare',
-            'category_type': 'health',
-            'amount': 150,
-            'period': 'monthly',
-            'include_subcategories': True
-        },
-        {
-            'name': 'Monthly Personal',
-            'category_type': 'personal',
-            'amount': 200,
-            'period': 'monthly',
-            'include_subcategories': True
-        },
-        {
-            'name': 'Weekly Grocery Budget',
-            'category_type': 'food',  # Will use subcategory if available
-            'amount': 150,
-            'period': 'weekly',
-            'include_subcategories': False,
-            'subcategory_name': 'Groceries'  # Try to find this subcategory
-        },
-        {
-            'name': 'Weekly Dining Out',
-            'category_type': 'food',  # Will use subcategory if available
-            'amount': 75,
-            'period': 'weekly',
-            'include_subcategories': False,
-            'subcategory_name': 'Restaurants'  # Try to find this subcategory
-        },
-        {
-            'name': 'Monthly Subscriptions',
-            'category_type': 'entertainment',  # Will use subcategory if available
-            'amount': 50,
-            'period': 'monthly',
-            'include_subcategories': False,
-            'subcategory_name': 'Subscriptions'  # Try to find this subcategory
-        },
-        {
-            'name': 'Annual Vacation',
-            'category_type': 'personal',
-            'amount': 1500,
-            'period': 'yearly',
-            'include_subcategories': False
-        }
-    ]
-    
-    # Add budgets to database
-    budgets_added = 0
-    
-    for budget_template in default_budgets:
-        # Determine the category ID to use
-        category_id = None
-        cat_type = budget_template['category_type']
-        
-        if cat_type in category_map:
-            category_id = category_map[cat_type]
-            
-            # Check for subcategory if specified
-            if 'subcategory_name' in budget_template:
-                # Find the category
-                main_category = Category.query.get(category_id)
-                if main_category and hasattr(main_category, 'subcategories'):
-                    # Look for matching subcategory
-                    for subcat in main_category.subcategories:
-                        if budget_template['subcategory_name'].lower() in subcat.name.lower():
-                            category_id = subcat.id
-                            break
-        
-        # If we have a valid category, create the budget
-        if category_id:
-            new_budget = Budget(
-                user_id=user_id,
-                category_id=category_id,
-                name=budget_template['name'],
-                amount=budget_template['amount'],
-                period=budget_template['period'],
-                include_subcategories=budget_template.get('include_subcategories', True),
-                start_date=datetime.utcnow(),
-                is_recurring=True,
-                active=False,  # Deactivated by default
-                created_at=datetime.utcnow(),
-                updated_at=datetime.utcnow()
-            )
-            
-            db.session.add(new_budget)
-            budgets_added += 1
-    
-    if budgets_added > 0:
-        db.session.commit()
-        
-    return budgets_added
 
 def update_currency_rates():
     """
@@ -1505,93 +1073,7 @@ def send_welcome_email(user):
         app.logger.error(f"Error sending welcome email: {str(e)}")
         return False
 
-def send_group_invitation_email(user, group, inviter):
-    """
-    Send an email notification when a user is added to a group
-    """
-    try:
-        subject = f"You've been added to {group.name} on Dollar Dollar Bill Y'all"
-        
-        # Create invitation email body
-        body_html = f"""
-        <html>
-        <head>
-            <style>
-                body {{ font-family: Arial, sans-serif; line-height: 1.6; color: #333; }}
-                .container {{ max-width: 600px; margin: 0 auto; padding: 20px; }}
-                .header {{ background-color: #15803d; color: white; padding: 10px 20px; text-align: center; border-radius: 5px 5px 0 0; }}
-                .content {{ background-color: #f9f9f9; padding: 20px; border-radius: 0 0 5px 5px; }}
-                .footer {{ text-align: center; margin-top: 20px; font-size: 12px; color: #777; }}
-                .button {{ display: inline-block; background-color: #15803d; color: white; padding: 10px 20px; text-decoration: none; border-radius: 5px; margin-top: 15px; }}
-            </style>
-        </head>
-        <body>
-            <div class="container">
-                <div class="header">
-                    <h1>Group Invitation</h1>
-                </div>
-                <div class="content">
-                    <h2>Hi {user.name},</h2>
-                    <p>You have been added to the group <strong>"{group.name}"</strong> by {inviter.name}.</p>
-                    
-                    <h3>Group Details:</h3>
-                    <ul>
-                        <li><strong>Group Name:</strong> {group.name}</li>
-                        <li><strong>Description:</strong> {group.description or 'No description provided'}</li>
-                        <li><strong>Created by:</strong> {group.creator.name}</li>
-                        <li><strong>Members:</strong> {', '.join([member.name for member in group.members])}</li>
-                    </ul>
-                    
-                    <p>You can now track and split expenses with other members of this group.</p>
-                    
-                    <a href="{request.host_url}groups/{group.id}" class="button">View Group</a>
-                </div>
-                <div class="footer">
-                    <p>This email was sent to {user.id}. If you believe this was a mistake, please contact the group creator.</p>
-                </div>
-            </div>
-        </body>
-        </html>
-        """
-        
-        # Simple text version for clients that don't support HTML
-        body_text = f"""
-        Group Invitation - Dollar Dollar Bill Y'all
-        
-        Hi {user.name},
-        
-        You have been added to the group "{group.name}" by {inviter.name}.
-        
-        Group Details:
-        - Group Name: {group.name}
-        - Description: {group.description or 'No description provided'}
-        - Created by: {group.creator.name}
-        - Members: {', '.join([member.name for member in group.members])}
-        
-        You can now track and split expenses with other members of this group.
-        
-        View Group: {request.host_url}groups/{group.id}
-        
-        This email was sent to {user.id}. If you believe this was a mistake, please contact the group creator.
-        """
-        
-        # Create and send the message
-        msg = Message(
-            subject=subject,
-            recipients=[user.id],
-            body=body_text,
-            html=body_html
-        )
-        
-        # Send the email
-        mail.send(msg)
-        
-        app.logger.info(f"Group invitation email sent to {user.id} for group {group.name}")
-        return True
-        
-    except Exception as e:
-        app.logger.error(f"Error sending group invitation email: {str(e)}")
-        return False
+
     
 
 @app.before_first_request
@@ -1763,20 +1245,6 @@ def utility_processor():
     }
     
 
-
-
-
-
-
-
-
-
-    
-
-
-
-
-
 @app.route('/get_category_splits/<int:expense_id>')
 @login_required_dev
 def get_category_splits(expense_id):
@@ -1837,169 +1305,7 @@ def get_category_splits(expense_id):
             'success': False,
             'message': f'Error: {str(e)}'
         }), 500
-#--------------------
-# ROUTES: GROUPS
-#--------------------
 
-@app.route('/groups')
-@login_required_dev
-def groups():
-    groups = Group.query.join(group_users).filter(group_users.c.user_id == current_user.id).all()
-    all_users = User.query.all()
-    return render_template('groups.html', groups=groups, users=all_users)
-
-@app.route('/groups/create', methods=['POST'])
-@login_required_dev
-def create_group():
-    try:
-        name = request.form.get('name')
-        description = request.form.get('description')
-        member_ids = request.form.getlist('members')
-        
-        group = Group(
-            name=name,
-            description=description,
-            created_by=current_user.id
-        )
-        
-        # Add creator as a member
-        group.members.append(current_user)
-        
-        # Add selected members
-        for member_id in member_ids:
-            user = User.query.filter_by(id=member_id).first()
-            if user and user != current_user:
-                group.members.append(user)
-        
-        db.session.add(group)
-        db.session.commit()
-        flash('Group created successfully!')
-    except Exception as e:
-        flash(f'Error creating group: {str(e)}')
-    
-    return redirect(url_for('groups'))
-
-@app.route('/groups/<int:group_id>')
-@login_required_dev
-def group_details(group_id):
-    base_currency = get_base_currency()
-    group = Group.query.get_or_404(group_id)
-
-    # Check if user is member of group
-    if current_user not in group.members:
-        flash('Access denied. You are not a member of this group.')
-        return redirect(url_for('groups'))
-    categories = Category.query.filter_by(user_id=current_user.id).order_by(Category.name).all()
-    
-    expenses = Expense.query.filter_by(group_id=group_id).order_by(Expense.date.desc()).all()
-    all_users = User.query.all()
-    currencies = Currency.query.all()
-    return render_template('group_details.html', 
-                           group=group, 
-                           expenses=expenses,
-                           currencies=currencies, 
-                           base_currency=base_currency,
-                           categories=categories,
-                           users=all_users)
-
-@app.route('/groups/<int:group_id>/add_member', methods=['POST'])
-@login_required_dev
-def add_group_member(group_id):
-    group = Group.query.get_or_404(group_id)
-    if current_user != group.creator:
-        flash('Only group creator can add members')
-        return redirect(url_for('group_details', group_id=group_id))
-    
-    member_id = request.form.get('user_id')
-    user = User.query.filter_by(id=member_id).first()
-    
-    if user and user not in group.members:
-        group.members.append(user)
-        db.session.commit()
-        flash(f'{user.name} added to group!')
-        create_default_categories(user.id)
-        # Send group invitation email
-        try:
-            send_group_invitation_email(user, group, current_user)
-        except Exception as e:
-            app.logger.error(f"Failed to send group invitation email: {str(e)}")
-    
-    return redirect(url_for('group_details', group_id=group_id))
-
-@app.route('/groups/<int:group_id>/remove_member/<member_id>', methods=['POST'])
-@login_required_dev
-def remove_group_member(group_id, member_id):
-    group = Group.query.get_or_404(group_id)
-    if current_user != group.creator:
-        flash('Only group creator can remove members')
-        return redirect(url_for('group_details', group_id=group_id))
-    
-    user = User.query.filter_by(id=member_id).first()
-    if user and user in group.members and user != group.creator:
-        group.members.remove(user)
-        db.session.commit()
-        flash(f'{user.name} removed from group!')
-    
-    return redirect(url_for('group_details', group_id=group_id))
-
-@app.route('/groups/<int:group_id>/delete', methods=['GET', 'POST'])
-@login_required_dev
-def delete_group(group_id):
-    """Delete a group and its associated expenses"""
-    # Find the group
-    group = Group.query.get_or_404(group_id)
-    
-    # Security check: Only the creator can delete the group
-    if current_user.id != group.created_by:
-        flash('Only the group creator can delete the group', 'error')
-        return redirect(url_for('group_details', group_id=group_id))
-    
-    # GET request shows confirmation prompt, POST actually deletes
-    if request.method == 'GET':
-        # Count associated expenses
-        expense_count = Expense.query.filter_by(group_id=group_id).count()
-        # Set up session data for confirmation
-        session['delete_group_id'] = group_id
-        session['delete_group_name'] = group.name
-        session['delete_group_expense_count'] = expense_count
-        # Show confirmation toast
-        flash(f'Warning: Deleting this group will also delete {expense_count} associated transactions. This action cannot be undone.', 'warning')
-        return redirect(url_for('group_details', group_id=group_id))
-    
-    # POST request (actual deletion)
-    try:
-        # Get stored values from session
-        group_name = session.get('delete_group_name', group.name)
-        expense_count = session.get('delete_group_expense_count', 0)
-        
-        # Delete associated expenses first
-        Expense.query.filter_by(group_id=group_id).delete()
-        
-        # Delete the group
-        db.session.delete(group)
-        db.session.commit()
-        
-        # Clear session data
-        session.pop('delete_group_id', None)
-        session.pop('delete_group_name', None)
-        session.pop('delete_group_expense_count', None)
-        
-        # Success message
-        if expense_count > 0:
-            flash(f'Group "{group_name}" and {expense_count} associated transactions have been deleted', 'success')
-        else:
-            flash(f'Group "{group_name}" has been deleted', 'success')
-            
-        return redirect(url_for('groups'))
-        
-    except Exception as e:
-        db.session.rollback()
-        app.logger.error(f"Error deleting group {group_id}: {str(e)}")
-        flash(f'Error deleting group: {str(e)}', 'error')
-        return redirect(url_for('group_details', group_id=group_id))
-#--------------------
-# ROUTES: ADMIN
-#--------------------
 
 @app.route('/admin')
 @login_required_dev
